dpif-netdev: Fix conn offload numbers
The offload provider does not deal with whole connections at once.
It only receives one side, then the other. Consequently, it counts each
side as one 'conn' object having been inserted.

So the 'n_conns' that we get from it has double the value, and should be
halved when translated into the 'ct_bidir' stat that dpif-netdev
reports.

Fixes: d36ea0eaad7b ("dpif-netdev: Use the new offload stats API")
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>