dpif-netdev: Query e2e-cache stats if MT query fails
With the new ct-offload-api, the offloaded connection stats is queried only
using their MT ufids.
e2e-cache flows ufids are different from their MT coutnerparts and require a
recalc of the ufid before query, this was previously done in
e2e_cache_get_merged_flows_stats() but the call to this function was removed
when we moved to ct-offload-api. Fix it.

Issue: OVS-3387634
Fixes: 60187817e ("dpif-netdev: Switch to CT offload API")
Signed-off-by: Salem Sol <salems@nvidia.com>