netdev-offload-dpdk: Fix packet reg recover condition check
A packet which has meta + mark is assumed to be CT, otherwise if only meta exists
it's assumed to be sflow. Due to a mistake in checking the return value the if
condition was wrong and caused wrong logic of assuming mark + meta is sflow
instead of CT. Fix it.

Issue: OVS-3405350
Fixes: 0f2d638c41 ("netdev-offload-dpdk: Abstract flow info action")
Signed-off-by: Salem Sol <salems@nvidia.com>