dpdk-offload-doca: Introduce doca eswitch context
The purpose of the new context is to allow creation
and management of resources that are global per eswitch
and shared among all the eswitch ports.

The infrastructure uses offload metadata api to reference
such contexts based on eswitch manager port id and therefore
all the ports that belong to the same eswitch will
reference and share the same context.

Signed-off-by: Ariel Levkovich <lariel@nvidia.com>