dpdk-offload-doca: Match on L4 in ct-zone rules
Add a match on the L4 protocol type in the default fixed rules in the CT
zones. To progress toward the CT tables, a packet must either be UDP or
TCP with the flags syn, rst, and fin unset.

Instead of a single rule matching that the CT state is *not visited*,
transform into two rules each matching the relevant L4 fields.

If neither of those match, a last catch-all miss rule is added, that
will trigger for unsupported L4 protocols and TCP control packets.

Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>