netdev-offload-dpdk: Fix vport counters
Counting offloads elements is broken when a vport is involved.
Three elements are counted:

  * Number of insertions (n_offloads)
  * Number of flows handled (n_flows)
  * Number of connections handled (n_conns)

'n_offloads' increments are done on the physdev right after a successful
insertion. Decrements however are done on the vport during removal.
Additionally, cleanups done after insertion failure won't be counted.

For 'n_offloads', consider that we want to know exactly how many rules
are currently within a specific port hardware, and put all counting on
the physdev.

For 'n_flows' and 'n_conns', the information is more abstract and
addressed to the user to know whether a port had the higher-level
objects it is related (megaflows and connections) offloaded. For them,
it makes sense to track on the vport, such that if we query e.g. a
VXLAN interface, we will know how many of its flows or connections
are currently offloaded.

For those two fields, modify the increments to be counted on the vport
and not the physdev, when applicable.

Fixes: bc125728135b ("netdev-offload-dpdk: Implement get_stats function")
Fixes: 0e6366c2399d ("netdev-offload-dpdk: Implement hw-offload statistics read.")
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>