acinclude: Check for missing doca-flow-lib earlier
Currently we check for -lWhole-archive before checking for the doca-flow.h,
in case we compiled doca and it didn't include doca-flow lib then the check
for pkg-config will result in an error and print out a misleading message. Fix
this by checking for doca-flow.h existing before validating the pkg-config version.

Issue: OVS-3409231
Signed-off-by: Salem Sol <salems@nvidia.com>