dpif-netdev: Add dpif-netdev/dump-packets appctl cmd
ovs-appctl dpif-netdev/dump-packets [on/off].
Default is off. If no parameter is provided, set to on.

When it is on, every packet going to SW is dumped to ovs log, in a
scapy format.

For example (lines are broken to fit commit message):

port_no=2: in_port=2, recirc_id=0, size=110.
    Raw(hex_bytes('b8cef605e73e946dae0ca71e08004500006000010000401199396fa
                   801026fa80101043212b5004cc44a08000000000064001aec6e961e
                   cbe4112233445508004500002e000100004011b6ff01010102'))/
    Raw('\x00' * 30)
port_no=5: in_port=5, recirc_id=0, size=60.
    Raw(hex_bytes('e411223344551aec6e961ecb08004500002e000100004011b6ffc0a
                   801140101010213891388001a00006162636465666768696a6b6c6d
                   6e6f707172'))

Signed-off-by: Eli Britstein <elibr@nvidia.com>