dpctl: Sort flows by port and recirc ID by default
The sorted output is much more readable and usable.
Using the 'sort=' option however is difficult, as you need to
remember the syntax and which fields are most useful.

Use a default sorting method, applied when no 'sort=' parameter is used.

Modify the tests to match the new order of the output.

Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>
Signed-off-by: Eli Britstein <elibr@nvidia.com>