netdev-offload-dpdk: Allow all actions in pre-ct
As the flow to execute can be found from the packet's mark (e.g. partial
offload), not depending on the packet content, as well as skipping the
actions already done by HW, we can allow changing the packet in the HW
pre-ct flow and still recover correctly.
Count all actions in pre_ct_cnt, not only output ones to be able to
correctly set skip_actions for pre-ct flows, and allow all actions in
pre-ct.

Issue: OVS-3415162
Signed-off-by: Eli Britstein <elibr@nvidia.com>