dpif-netdev: Change mark to flow map and offload ordering
Allowing packet modification actions before CT is only possible
if the miss finds the correct flow according to its mark, otherwise
an incorrect flow will be created.
Currently, the mapping is created only after the offload is created, and
destroyed before the offload is destroyed. A race condition may occur in
case a packet hits the HW flow, but its mark is not in the map yet.
Change the ordering to cover for such case.

Issue: OVS-3415162
Signed-off-by: Eli Britstein <elibr@nvidia.com>