offload-metadata: Do not lookup entries again before removal
When removing entries from the maps, an ultimate lookup is performed
to ensure that the entry is still within the map.

This ultimate lookup is made unnecessary by the use of refcounts. If we
pass through the very last entry unref, we are sure to be the only
thread possibly removing this entry.

Furthermore, all entries being passed to the function are the actual
elements of which their 'cmap_node' is currently within the maps. There
is no need to iterate to find it back.

This change improves the performance of 'free' operations:

Running with N=1,2,4

  ./utilities/bench.py run -- \
  ./tests/ovstest test-offload-metadata benchmark -csv 1000000 $N > after.$N

  ./utilities/bench.py compare {before,after}.$N

                             N=1     N=2     N=4
      id:0.0-alloc: mean   +0.1%   -0.2%   +0.3%
       id:0.1-free: mean   -9.0%   -7.9%   -8.5%
      id:1.0-alloc: mean   -0.4%   +2.1%   -0.7%
        id:1.1-ref: mean   +0.1%   -0.6%   -1.5%
      id:1.2-unref: mean   +2.1%   -2.2%   +6.4%
       id:1.3-free: mean   -7.4%   -7.8%   -8.5%
      id:2.0-mixed: mean   +2.1%   -1.7%   -0.7%
       id:2.1-free: mean   +0.6%   -5.7%   -4.2%
    priv:3.0-alloc: mean   +2.2%   -2.4%   +3.0%
     priv:3.1-free: mean  -13.5%  -15.5%   -9.3%
    priv:4.0-alloc: mean   +5.4%   -1.7%   +3.3%
      priv:4.1-ref: mean   +3.4%   -2.4%   -1.5%
    priv:4.2-unref: mean  +21.3%  -11.0%  +12.0%
     priv:4.3-free: mean  -12.6%  -16.8%   -9.1%
    priv:5.0-mixed: mean   +2.1%   -5.6%   +2.8%
     priv:5.1-free: mean   -8.6%  -12.4%   -4.7%
    priv:6.0-alloc: mean   +4.6%   -5.4%   +2.0%
priv:6.1-pos-query: mean   +3.2%   -2.0%   +1.7%
priv:6.2-neg-query: mean   +3.7%   -1.9%   -0.8%
     priv:6.3-free: mean  -13.3%  -16.0%  -10.6%
priv:6.4-neg-query: mean   +2.5%   -1.6%   +1.5%

Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>