netdev-offload-dpdk: Aux tables init/uninit from main thread
Previously, aux tables was initialized in a lazy mode, upon the first
flow offload for a netdev, in the context of the the first offload thread
for this netdev. The flush is a message sent to all offload threads. The
creating thread id was kept in the aux flows, and destruction was
ignored by other threads than the one created.
As we give up counting statistics for those flows, that must be from an
offload thread, those init/uninit are moved to the netdev init/destroy,
not in a lazy mode.
This will shorten the delay introduced by this init in the first flows.

Signed-off-by: Eli Britstein <elibr@nvidia.com>