utilities: Add bench.py script
Add a uility script to help analyze the results of module benchmarks.
A 'runner' mode executes commands given in parameters and consume
their CSV output for analysis.

Storing that output into files, it is then possible to use the
script to compare multiple outputs, and show the evolution
of each result lines.

Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>