netdev-offload-dpdk: Retry port-add upon aux-init failure
As part of assigning flow_api, there is a call to aux-init, that should
configure flows on the ESW. To configure flows on representors, DPDK
requires the ESW to be already started.
If the port-add does not fail, the aux-init will fail. After [1], there
will not be any retries.

In such case, propagate EAGAIN to fail the port-add. Once the ESW port
is added, a "reconfigure" is done automatically by OVS, that will retry
to do the port-adds for the other ports. This time, the aux-init will
succeed.

Also, in case of a failure, rollback memory allocations etc to fix a
memory leak.

[1]: f292a241348c ("netdev-offload-dpdk: Aux tables init/uninit from main thread")

Fixes: f292a241348c ("netdev-offload-dpdk: Aux tables init/uninit from main thread")
Signed-off-by: Eli Britstein <elibr@nvidia.com>