netdev-offload-dpdk: Pre-allocate flow handles
In async mode, we must post a handle description to the insertion queue,
for the completion callback to fill in with the actual entry pointer
once completed.

This can only be done if the handle provided is the final one, that will
be stored in the port offload handle map.

Instead of using stack memory for insertion then copying the handles
on success, immediately allocate the entry and use it.

Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>