offload-metadata: Fix delayed priv uninit not being executed
If a priv reference is taken twice and release twice within a delay
period of the offload-metadata, the priv_uninit callback will not be
executed. This is because the unref is done during the entry release,
not when posting the delayed_release item.

During the insertion of the delayed_release item, the refcount
will be seen at '2' in the example above.

Change the logic, and execute the priv_uninit only after actually
unreferencing. There is no other way to correctly write this.

Add relevant verifications in the unit-test to detect that bug
and ensure it remains fixed.

Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>