dpdk-offload-doca: Merge CT pipes
Collapse the three different CT pipes into a single one per L4 protocol.
Use the 5 possible CT actions into a single template within that pipe.

As a result, the five fields of the 5-tuple are always rewritten by CT.
If a NAT action is part of the connection, the translated fields
are modified, otherwise the original value of the tuple is written.

It reduces the amount of memory consumed by basic pipes three-fold.

Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>