offload-metadata: Iterate on delayed list properly
Use the better macro LIST_FOR_EACH_SAFE to iterate on the delayed item
list. It is easier to read and simpler to implement.

Additionally, the iteration is much faster, as the test
'ovs_list_is_empty()' happens much less often. It completely makes that
function disappear from profiling on bf2.

Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>