dpdk-offload-doca: Implement insert_conn API
CT offload over doca is done with basic pipes with pre-set templates.
Other fields' values not in a template are don't care, not in match
template neither in action template.
For example, action template for PLAIN CT does not perform header rewrite,
so even setting such values for that pipe is ok.
Other pre-set parameters, such as masks, forward, count do not have to be
provided, but to be taken from the pipe configuration.

Implement insert_conn API, taking advantage of the above, to translate
ct-offload properties directly to create doca flows for connections w/o
not needd rte logic.

Signed-off-by: Eli Britstein <elibr@nvidia.com>