Makefile.am: Enable DOCA experimental API always
Fix how linking is evaluated for doca-flow lib.
In particular, on systems that do include the warning
'-Wno-deprecated-declarations' by default, the macro
'DOCA_ALLOW_EXPERIMENTAL_API' is required, and it is not
provided by the 'pkg-config --cflags doca'.

Automatically enable experimental DOCA symbols.
All DOCA-flow is experimental, which is what we use from DOCA.

Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>