dpdk-offload-doca: Introduce hash pipe ctx
The action to be supported is a combination of hash(),recirc(). As a
pre-step towards it, have a hash pipe ctx. Only IPv4 traffic is
supported. Classify to TCP/UDP/simple IPv4-L3.
Each hash pipe and its entry copy the hash result to meta bits, enabling
them to go to SW for packet recovery, as well as future matching in HW.

Issue: OVS-3407068
Signed-off-by: Eli Britstein <elibr@nvidia.com>