dpif-netdev: Fix mark to flow disassociation in case of failure
Without this patch for cases of a flow modify we will call dp_netdev_flow_offload_put()
and if the offload of the modified flow fails the old offload is deleted and we
will never free the SW mark. Fix it by always disassociating the mark from a
flow in cases of failure.

Issue: OVS-3494447
Fixes: 6ec9708c4482 ("dpif-netdev: Change mark to flow map and offload ordering")
Signed-off-by: Salem Sol <salems@nvidia.com>