offload-metadata: Fix delayed priv uninit
Delayed init and uninit in offload metadata is only used for managing
flow miss contexts. The miss rule of these MD must be immediately
removed while the entry itself must have a delay.

The previous fix to offload-metadata missing priv_uninit only fixed
the offload-metadata logic in a vacuum. It breaks the flow miss context
handling.

Revert back to a mode where the priv is uninitialized immediately,
regardless of a delay being set.

To prevent the double reference preventing sometimes priv uninit from
executing, separate the reference counting between the entry itself
which is delayed and the priv which is not.

It allows triggering the uninit at the immediate release call,
while allowing to take back a reference in the delayed mode.

Fixes: 6e92cd562784 ("offload-metadata: Fix delayed priv uninit not being executed")
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>