offload-metadata: Fix associated data priv references
In 'associated' mode, there is no priv init or reference taking
when creating the shallow reference to the original node.

When destroying an entry, only dereference the priv if
the node is not an associated one.

Add a unit test verifying offload-metadata operations when
using ID association and priv init/uninit at the same time.

This is used currently by the flow-id <-> mark map but it
is not tested.

Fixes: 8a3f9c5d598d ("offload-metadata: Fix delayed priv uninit")
Signed-off-by: Gaetan Rivet <gaetanr@nvidia.com>