netdev-offload-dpdk: Direct CT access to offload_data
As the offload_data (struct ufid_to_rte_flow_data) is provided in
struct ct_flow_offload_item, there is no need to find it by its ufid.
Use it instead.

The mapping by ufid is still in place, so in a case of flushing the
netdev, the objects can be properly destroyed.

Signed-off-by: Eli Britstein <elibr@nvidia.com>