netdev-dpdk: Aligh meter ID with the user provided value
OpenFlow standard specifies meter IDs starting from ID 1. Internally,
ovs ofproto will offset the ID by -1 before passing it to
dpif_offload_class meter methods. This way, the internal meter array
can start with index 0.

On the other hand, such approach introduces an inconvenience when
passing a meter ID around as uint32_t value. There is no way to tell if
the meter ID has been set or not. So the code, that handles meter
offloading, used to offset the IDs by +1 when passing them around, and
later compensate for that by -1 when referring to the actual meter ID.
That is an error prone approach, susceptible to off by one errors.

To simplify meter IDs handling all the dpif_offload_class methods, that
are entry points for meter offloading, adjust meter IDs by +1. So the
meter ID 0 becomes an invalid value, which makes it trivial to test if
the meter ID has been set or not.

Signed-off-by: Dima Chumak <dchumak@nvidia.com>