dpdk-offload-doca: Add global post meter table
Introduce post meter table that is used for meter action offloading in
the later patches in the series. The table is global and initialized at
eswitch init stage.

There is a single static rule for dropping packets that have RED meter
color in their metadata. The rules for GREEN color handling will be
inserted dynamically when a flow, that has a meter action, is offloaded.
The static red rule has a higher priority (prio 0) than green rules
(prio 1).

Signed-off-by: Dima Chumak <dchumak@nvidia.com>