netdev-dpdk: Ask for restart OpenvSwitch for settings changes requiring netdev reconfigure
Some settings changes like MTU require netdev reconfiguration, which will
stop the DPDK port. In order to stop DPDK port when working with HWS config
we must flush any inserted offloads, which should be done at the dpif-netdev
layer. As a temporary workaround for DOCA, block the netdev_reconfigure
to prevent breaking OVS and declare that it requires a restart.

Issue: OVS-3511377
Signed-off-by: Emeel Hakim <ehakim@nvidiacom>