netdev-offload-dpdk: Fix recirc after tunnel-push
The cited commit didn't take into consideration a scenario that there is
a terminating physical port and a recirc action. For example in CT and
mirroring to a tunnel:
actions: clone(tnl_push(...,out_port(br)),output(phys)),ct(),recirc()

In this case, the miss context has a wrong vport value.

Fix it by resetting the temp value of tnl_push_out_port in case of
output action.

Issue: OVS-3525807
Fixes: a8c2c6a50e7e ("netdev-offload-dpdk: Support recirc after tunnel-push")
Signed-off-by: Eli Britstein <elibr@nvidia.com>